gateway/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.8
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
gunicorn